+++ examples/subcommand.py
@@ -3,7 +3,10 @@ from pydantic import BaseModel
 
 
 class Clone(BaseModel):
-    """Clone a repo."""
+    """Clone a repo.
+
+    This is a subcommand.
+    """
 
     repo: str
     """Clone this repo."""
@@ -14,10 +17,13 @@ class Clone(BaseModel):
 
 
 class Merge(BaseModel):
-    """Merge a branch."""
+    """Merge a branch.
+
+    Provide a branch and merge it into your active branch.
+    """
 
     branch: str
-    """Branch to merge into active branch."""
+    """Branch name."""
 
     def main(self):
         """Run when this subcommand is choosen."""