+++ tests/test_choice.py
@@ -41,11 +41,19 @@ def test_optional_none_choice():
 
 
 def test_choice_help(capture_output):
-    with pytest.raises(SystemExit):
+    with pytest.raises(SystemExit) as err:
         capture_output(ModelWithChoice, ["-h"])
 
-    assert "Arguments" in capture_output.captured_output
-    assert "[allowed values: option1, option2]" in capture_output.captured_output
+    assert err.value.code == 0
+    assert (
+        """
+Usage: my-cli-app [Arguments]
+
+Arguments:
+    choice                [allowed values: option1, option2]
+"""
+        == capture_output.captured_output
+    )
 
 
 def test_optional_choice_help(capture_output):
@@ -73,7 +81,7 @@ def test_optional_none_choice_help(capture_output):
 Usage: my-cli-app [Options]
 
 Options:
-    --choice             A choice with a default. [Optional] [default = None]
+    --choice             A choice with a default. [allowed values: option1, option2] [default = None]
 """
         == capture_output.captured_output
     )
@@ -90,6 +98,18 @@ Usage: my-cli-app [Options]
 
 Options:
     --choice             A choice with a default. [allowed values: 1, 2] [default = 1]
+"""
+        == capture_output.captured_output
+    )
+
+
+def test_failing_choice(capture_output):
+    with pytest.raises(SystemExit) as err:
+        capture_output(ModelWithOptionalChoice, ["--choice", "option3"])
+
+    assert err.value.code == 1
+    assert (
+        """ERROR: Incorrect value for --choice (option3). Input should be 'option1' or 'option2'
 """
         == capture_output.captured_output
     )